massive refactor: cut out old constraints entirely
