+++ jolt-core/src/zkvm/r1cs/constraints.rs
@@ -1,282 +1,332 @@
-use crate::{field::JoltField, zkvm::instruction::CircuitFlags};
+//! Compile-time constant R1CS constraints
+//!
+//! This module provides a static, compile-time representation of R1CS constraints
+//! to replace the dynamic constraint building in the prover's hot path.
 
-use super::{
-    builder::{CombinedUniformBuilder, R1CSBuilder},
-    inputs::JoltR1CSInputs,
-};
+use super::inputs::JoltR1CSInputs;
+use crate::field::JoltField;
+use crate::poly::multilinear_polynomial::MultilinearPolynomial;
+use crate::zkvm::instruction::CircuitFlags;
 
-pub const PC_START_ADDRESS: i64 = 0x80000000;
+// Re-export key types from ops module for convenience
+pub use super::ops::{ConstLC, ConstTerm};
 
-pub trait R1CSConstraints<F: JoltField> {
-    fn construct_constraints(padded_trace_length: usize) -> CombinedUniformBuilder<F> {
-        let mut uniform_builder = R1CSBuilder::new();
-        Self::uniform_constraints(&mut uniform_builder);
-
-        CombinedUniformBuilder::construct(uniform_builder, padded_trace_length)
-    }
-    /// Constructs Jolt's uniform constraints.
-    /// Uniform constraints are constraints that hold for each step of
-    /// the execution trace.
-    fn uniform_constraints(builder: &mut R1CSBuilder);
+/// r1cs_eq_conditional!: verbose, condition-first equality constraint
+/// Usage: r1cs_eq_conditional!(if { COND } => { LEFT } == { RIGHT });
+#[macro_export]
+macro_rules! r1cs_eq_conditional {
+    (if { $($cond:tt)* } => ( $($left:tt)* ) == ( $($right:tt)* ) ) => {{
+        $crate::zkvm::r1cs::constraints::constraint_eq_conditional_lc(
+            $crate::lc!($($cond)*),
+            $crate::lc!($($left)*),
+            $crate::lc!($($right)*),
+        )
+    }};
 }
 
-pub struct JoltRV32IMConstraints;
-impl<F: JoltField> R1CSConstraints<F> for JoltRV32IMConstraints {
-    fn uniform_constraints(cs: &mut R1CSBuilder) {
-        // if LeftOperandIsRs1Value { assert!(LeftInstructionInput == Rs1Value) }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsRs1Value),
-            JoltR1CSInputs::LeftInstructionInput,
-            JoltR1CSInputs::Rs1Value,
-        );
+/// r1cs_if_else!: verbose if-then-else with explicit result
+/// Usage: r1cs_if_else!(if { COND } => { TRUE } else { FALSE } => { RESULT });
+#[macro_export]
+macro_rules! r1cs_if_else {
+    (if { $($cond:tt)* } => ( $($tval:tt)* ) else ( $($fval:tt)* ) => ( $($result:tt)* ) ) => {{
+        $crate::zkvm::r1cs::constraints::constraint_if_else_lc(
+            $crate::lc!($($cond)*),
+            $crate::lc!($($tval)*),
+            $crate::lc!($($fval)*),
+            $crate::lc!($($result)*),
+        )
+    }};
+}
 
-        // if LeftOperandIsPC { assert!(LeftInstructionInput == UnexpandedPC) }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsPC),
-            JoltR1CSInputs::LeftInstructionInput,
-            JoltR1CSInputs::UnexpandedPC,
-        );
+/// r1cs_prod!: product constraint
+/// Usage: r1cs_prod!({ LEFT } * { RIGHT } == { RESULT });
+#[macro_export]
+macro_rules! r1cs_prod {
+    ( ( $($left:tt)* ) * ( $($right:tt)* ) == ( $($result:tt)* ) ) => {{
+        $crate::zkvm::r1cs::constraints::constraint_prod_lc(
+            $crate::lc!($($left)*),
+            $crate::lc!($($right)*),
+            $crate::lc!($($result)*),
+        )
+    }};
+}
 
-        // if !(LeftOperandIsRs1Value || LeftOperandIsPC)  {
-        //     assert!(LeftInstructionInput == 0)
-        // }
-        // Note that LeftOperandIsRs1Value and LeftOperandIsPC are mutually exclusive flags
-        cs.constrain_eq_conditional(
-            1 - JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsRs1Value)
-                - JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsPC),
-            JoltR1CSInputs::LeftInstructionInput,
-            0i128,
-        );
+/// A single R1CS constraint row
+#[derive(Clone, Copy, Debug)]
+pub struct ConstraintConst {
+    pub a: ConstLC,
+    pub b: ConstLC,
+    pub c: ConstLC,
+}
 
-        // if RightOperandIsRs2Value { assert!(RightInstructionInput == Rs2Value) }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsRs2Value),
-            JoltR1CSInputs::RightInstructionInput,
-            JoltR1CSInputs::Rs2Value,
-        );
+impl ConstraintConst {
+    pub const fn new(a: ConstLC, b: ConstLC, c: ConstLC) -> Self {
+        Self { a, b, c }
+    }
 
-        // if RightOperandIsImm { assert!(RightInstructionInput == Imm) }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsImm),
-            JoltR1CSInputs::RightInstructionInput,
-            JoltR1CSInputs::Imm,
-        );
-
-        // if !(RightOperandIsRs2Value || RightOperandIsImm)  {
-        //     assert!(RightInstructionInput == 0)
-        // }
-        // Note that RightOperandIsRs2Value and RightOperandIsImm are mutually exclusive flags
-        cs.constrain_eq_conditional(
-            1 - JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsRs2Value)
-                - JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsImm),
-            JoltR1CSInputs::RightInstructionInput,
-            0i128,
-        );
-
-        // if Load || Store {
-        //     assert!(RamAddress == Rs1Value + Imm)
-        // } else {
-        //     assert!(RamAddress == 0)
-        // }
-        let is_load_or_store = JoltR1CSInputs::OpFlags(CircuitFlags::Load)
-            + JoltR1CSInputs::OpFlags(CircuitFlags::Store);
-        cs.constrain_if_else(
-            is_load_or_store,
-            JoltR1CSInputs::Rs1Value + JoltR1CSInputs::Imm,
-            0i128,
-            JoltR1CSInputs::RamAddress,
-        );
-
-        // if Load {
-        //     assert!(RamReadValue == RamWriteValue)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Load),
-            JoltR1CSInputs::RamReadValue,
-            JoltR1CSInputs::RamWriteValue,
-        );
-
-        // if Load {
-        //     assert!(RamReadValue == RdWriteValue)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Load),
-            JoltR1CSInputs::RamReadValue,
-            JoltR1CSInputs::RdWriteValue,
-        );
-
-        // if Store {
-        //     assert!(Rs2Value == RamWriteValue)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Store),
-            JoltR1CSInputs::Rs2Value,
-            JoltR1CSInputs::RamWriteValue,
-        );
-
-        // if AddOperands || SubtractOperands || MultiplyOperands {
-        //     // Lookup query is just RightLookupOperand
-        //     assert!(LeftLookupOperand == 0)
-        // } else {
-        //     assert!(LeftLookupOperand == LeftInstructionInput)
-        // }
-        cs.constrain_if_else(
-            JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands)
-                + JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands)
-                + JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands),
-            0i128,
-            JoltR1CSInputs::LeftInstructionInput,
-            JoltR1CSInputs::LeftLookupOperand,
-        );
-
-        // If AddOperands {
-        //     assert!(RightLookupOperand == LeftInstructionInput + RightInstructionInput)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands),
-            JoltR1CSInputs::RightLookupOperand,
-            JoltR1CSInputs::LeftInstructionInput + JoltR1CSInputs::RightInstructionInput,
-        );
-
-        // If SubtractOperands {
-        //     assert!(RightLookupOperand == LeftInstructionInput - RightInstructionInput)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands),
-            JoltR1CSInputs::RightLookupOperand,
-            // Converts from unsigned to twos-complement representation
-            JoltR1CSInputs::LeftInstructionInput - JoltR1CSInputs::RightInstructionInput
-                + (0xffffffffffffffffi128 + 1),
-        );
-
-        // if MultiplyOperands {
-        //     assert!(RightLookupOperand == Rs1Value * Rs2Value)
-        // }
-        cs.constrain_prod(
-            JoltR1CSInputs::RightInstructionInput,
-            JoltR1CSInputs::LeftInstructionInput,
-            JoltR1CSInputs::Product,
-        );
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands),
-            JoltR1CSInputs::RightLookupOperand,
-            JoltR1CSInputs::Product,
-        );
-
-        // if !(AddOperands || SubtractOperands || MultiplyOperands || Advice) {
-        //     assert!(RightLookupOperand == RightInstructionInput)
-        // }
-        cs.constrain_eq_conditional(
-            1 - JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands)
-                - JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands)
-                - JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands)
-                // Arbitrary untrusted advice goes in right lookup operand
-                - JoltR1CSInputs::OpFlags(CircuitFlags::Advice),
-            JoltR1CSInputs::RightLookupOperand,
-            JoltR1CSInputs::RightInstructionInput,
-        );
-
-        // if Assert {
-        //     assert!(LookupOutput == 1)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Assert),
-            JoltR1CSInputs::LookupOutput,
-            1i128,
-        );
-
-        // if Rd != 0 && WriteLookupOutputToRD {
-        //     assert!(RdWriteValue == LookupOutput)
-        // }
-        cs.constrain_prod(
-            JoltR1CSInputs::Rd,
-            JoltR1CSInputs::OpFlags(CircuitFlags::WriteLookupOutputToRD),
-            JoltR1CSInputs::WriteLookupOutputToRD,
-        );
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::WriteLookupOutputToRD,
-            JoltR1CSInputs::RdWriteValue,
-            JoltR1CSInputs::LookupOutput,
-        );
-
-        // if Rd != 0 && Jump {
-        //     if !isCompressed {
-        //          assert!(RdWriteValue == UnexpandedPC + 4)
-        //     } else {
-        //          assert!(RdWriteValue == UnexpandedPC + 2)
-        //     }
-        // }
-        cs.constrain_prod(
-            JoltR1CSInputs::Rd,
-            JoltR1CSInputs::OpFlags(CircuitFlags::Jump),
-            JoltR1CSInputs::WritePCtoRD,
-        );
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::WritePCtoRD,
-            JoltR1CSInputs::RdWriteValue,
-            JoltR1CSInputs::UnexpandedPC + 4i128
-                - 2 * JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed),
-        );
-
-        // if Jump && !NextIsNoop {
-        //     assert!(NextUnexpandedPC == LookupOutput)
-        // }
-        cs.constrain_prod(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Jump),
-            1 - JoltR1CSInputs::NextIsNoop,
-            JoltR1CSInputs::ShouldJump,
-        );
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::ShouldJump,
-            JoltR1CSInputs::NextUnexpandedPC,
-            JoltR1CSInputs::LookupOutput,
-        );
-
-        // if Branch && LookupOutput {
-        //     assert!(NextUnexpandedPC == UnexpandedPC + Imm)
-        // }
-        cs.constrain_prod(
-            JoltR1CSInputs::OpFlags(CircuitFlags::Branch),
-            JoltR1CSInputs::LookupOutput,
-            JoltR1CSInputs::ShouldBranch,
-        );
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::ShouldBranch,
-            JoltR1CSInputs::NextUnexpandedPC,
-            JoltR1CSInputs::UnexpandedPC + JoltR1CSInputs::Imm,
-        );
-
-        // if !(ShouldBranch || Jump) {
-        //     if DoNotUpdatePC {
-        //         assert!(NextUnexpandedPC == UnexpandedPC)
-        //     } else if isCompressed {
-        //         assert!(NextUnexpandedPC == UnexpandedPC + 2)
-        //     } else {
-        //         assert!(NextUnexpandedPC == UnexpandedPC + 4)
-        //     }
-        // }
-        // Note that ShouldBranch and Jump instructions are mutually exclusive
-        cs.constrain_prod(
-            JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed),
-            JoltR1CSInputs::OpFlags(CircuitFlags::DoNotUpdateUnexpandedPC),
-            JoltR1CSInputs::CompressedDoNotUpdateUnexpPC,
-        );
-        cs.constrain_eq_conditional(
-            1 - JoltR1CSInputs::ShouldBranch - JoltR1CSInputs::OpFlags(CircuitFlags::Jump),
-            JoltR1CSInputs::NextUnexpandedPC,
-            JoltR1CSInputs::UnexpandedPC + 4i128
-                - 4 * JoltR1CSInputs::OpFlags(CircuitFlags::DoNotUpdateUnexpandedPC)
-                - 2 * JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed)
-                + 2 * JoltR1CSInputs::CompressedDoNotUpdateUnexpPC,
-        );
-
-        // if Inline {
-        //     assert!(NextPC == PC + 1)
-        // }
-        cs.constrain_eq_conditional(
-            JoltR1CSInputs::OpFlags(CircuitFlags::InlineSequenceInstruction),
-            JoltR1CSInputs::NextPC,
-            JoltR1CSInputs::PC + 1i128,
-        );
+    /// Evaluate this constraint at a specific row in the witness polynomials
+    /// Returns (a_eval, b_eval, c_eval) tuple
+    #[inline]
+    pub fn evaluate_row<F: JoltField>(
+        &self,
+        flattened_polynomials: &[MultilinearPolynomial<F>],
+        row: usize,
+    ) -> (F, F, F) {
+        let a_eval = self.a.evaluate_row(flattened_polynomials, row);
+        let b_eval = self.b.evaluate_row(flattened_polynomials, row);
+        let c_eval = self.c.evaluate_row(flattened_polynomials, row);
+        (a_eval, b_eval, c_eval)
     }
 }
+
+/// Number of uniform R1CS constraints
+pub const NUM_R1CS_CONSTRAINTS: usize = 28;
+
+// =============================================================================
+// CONSTRAINT BUILDER FUNCTIONS
+// =============================================================================
+/// Creates: condition * (left - right) == 0
+pub const fn constraint_eq_conditional_lc(
+    condition: ConstLC,
+    left: ConstLC,
+    right: ConstLC,
+) -> ConstraintConst {
+    ConstraintConst::new(
+        condition,
+        match left.add_const_lc(right.mul_by_const(-1)) {
+            Some(b) => b,
+            None => ConstLC::zero(),
+        },
+        ConstLC::zero(),
+    )
+}
+
+/// Creates: left * right == result
+pub const fn constraint_prod_lc(left: ConstLC, right: ConstLC, result: ConstLC) -> ConstraintConst {
+    ConstraintConst::new(left, right, result)
+}
+
+/// Creates: condition * (true_val - false_val) == (result - false_val)
+pub const fn constraint_if_else_lc(
+    condition: ConstLC,
+    true_val: ConstLC,
+    false_val: ConstLC,
+    result: ConstLC,
+) -> ConstraintConst {
+    ConstraintConst::new(
+        condition,
+        match true_val.add_const_lc(false_val.mul_by_const(-1)) {
+            Some(b) => b,
+            None => ConstLC::zero(),
+        },
+        match result.add_const_lc(false_val.mul_by_const(-1)) {
+            Some(c) => c,
+            None => ConstLC::zero(),
+        },
+    )
+}
+
+/// Static table of all 28 R1CS uniform constraints.
+pub static UNIFORM_R1CS: [ConstraintConst; NUM_R1CS_CONSTRAINTS] = [
+    // if LeftOperandIsRs1Value { assert!(LeftInstructionInput == Rs1Value) }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsRs1Value) } }
+        => ( { JoltR1CSInputs::LeftInstructionInput } ) == ( { JoltR1CSInputs::Rs1Value } )
+    ),
+    // if LeftOperandIsPC { assert!(LeftInstructionInput == UnexpandedPC) }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsPC) } }
+        => ( { JoltR1CSInputs::LeftInstructionInput } ) == ( { JoltR1CSInputs::UnexpandedPC } )
+    ),
+    // if !(LeftOperandIsRs1Value || LeftOperandIsPC)  {
+    //     assert!(LeftInstructionInput == 0)
+    // }
+    // Note that LeftOperandIsRs1Value and LeftOperandIsPC are mutually exclusive flags
+    r1cs_eq_conditional!(
+        if { { 1i128 } - { JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsRs1Value) } - { JoltR1CSInputs::OpFlags(CircuitFlags::LeftOperandIsPC) } }
+        => ( { JoltR1CSInputs::LeftInstructionInput } ) == ( { 0i128 } )
+    ),
+    // if RightOperandIsRs2Value { assert!(RightInstructionInput == Rs2Value) }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsRs2Value) } }
+        => ( { JoltR1CSInputs::RightInstructionInput } ) == ( { JoltR1CSInputs::Rs2Value } )
+    ),
+    // if RightOperandIsImm { assert!(RightInstructionInput == Imm) }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsImm) } }
+        => ( { JoltR1CSInputs::RightInstructionInput } ) == ( { JoltR1CSInputs::Imm } )
+    ),
+    // if !(RightOperandIsRs2Value || RightOperandIsImm)  {
+    //     assert!(RightInstructionInput == 0)
+    // }
+    // Note that RightOperandIsRs2Value and RightOperandIsImm are mutually exclusive flags
+    r1cs_eq_conditional!(
+        if { { 1i128 } - { JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsRs2Value) } - { JoltR1CSInputs::OpFlags(CircuitFlags::RightOperandIsImm) } }
+        => ( { JoltR1CSInputs::RightInstructionInput } ) == ( { 0i128 } )
+    ),
+    // if Load || Store {
+    //     assert!(RamAddress == Rs1Value + Imm)
+    // } else {
+    //     assert!(RamAddress == 0)
+    // }
+    r1cs_if_else!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::Load) } + { JoltR1CSInputs::OpFlags(CircuitFlags::Store) } }
+        => ( { JoltR1CSInputs::Rs1Value } + { JoltR1CSInputs::Imm } )
+        else ( { 0i128 } )
+        => ( { JoltR1CSInputs::RamAddress } )
+    ),
+    // if Load {
+    //     assert!(RamReadValue == RamWriteValue)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::Load) } }
+        => ( { JoltR1CSInputs::RamReadValue } ) == ( { JoltR1CSInputs::RamWriteValue } )
+    ),
+    // if Load {
+    //     assert!(RamReadValue == RdWriteValue)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::Load) } }
+        => ( { JoltR1CSInputs::RamReadValue } ) == ( { JoltR1CSInputs::RdWriteValue } )
+    ),
+    // if Store {
+    //     assert!(Rs2Value == RamWriteValue)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::Store) } }
+        => ( { JoltR1CSInputs::Rs2Value } ) == ( { JoltR1CSInputs::RamWriteValue } )
+    ),
+    // if AddOperands || SubtractOperands || MultiplyOperands {
+    //     // Lookup query is just RightLookupOperand
+    //     assert!(LeftLookupOperand == 0)
+    // } else {
+    //     assert!(LeftLookupOperand == LeftInstructionInput)
+    // }
+    r1cs_if_else!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands) } + { JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands) } + { JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands) } }
+        => ( { 0i128 } )
+        else ( { JoltR1CSInputs::LeftInstructionInput } )
+        => ( { JoltR1CSInputs::LeftLookupOperand } )
+    ),
+    // If AddOperands {
+    //     assert!(RightLookupOperand == LeftInstructionInput + RightInstructionInput)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands) } }
+        => ( { JoltR1CSInputs::RightLookupOperand } ) == ( { JoltR1CSInputs::LeftInstructionInput } + { JoltR1CSInputs::RightInstructionInput } )
+    ),
+    // If SubtractOperands {
+    //     assert!(RightLookupOperand == LeftInstructionInput - RightInstructionInput)
+    // }
+    // Converts from unsigned to twos-complement representation
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands) } }
+        => ( { JoltR1CSInputs::RightLookupOperand } ) == ( { JoltR1CSInputs::LeftInstructionInput } - { JoltR1CSInputs::RightInstructionInput } + { 0x10000000000000000i128 } )
+    ),
+    // if MultiplyOperands {
+    //     assert!(RightLookupOperand == Rs1Value * Rs2Value)
+    // }
+    r1cs_prod!(
+        ({ JoltR1CSInputs::RightInstructionInput }) * ({ JoltR1CSInputs::LeftInstructionInput })
+            == ({ JoltR1CSInputs::Product })
+    ),
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands) } }
+        => ( { JoltR1CSInputs::RightLookupOperand } ) == ( { JoltR1CSInputs::Product } )
+    ),
+    // if !(AddOperands || SubtractOperands || MultiplyOperands || Advice) {
+    //     assert!(RightLookupOperand == RightInstructionInput)
+    // }
+    // Arbitrary untrusted advice goes in right lookup operand
+    r1cs_eq_conditional!(
+        if { { 1i128 } - { JoltR1CSInputs::OpFlags(CircuitFlags::AddOperands) } - { JoltR1CSInputs::OpFlags(CircuitFlags::SubtractOperands) } - { JoltR1CSInputs::OpFlags(CircuitFlags::MultiplyOperands) } - { JoltR1CSInputs::OpFlags(CircuitFlags::Advice) } }
+        => ( { JoltR1CSInputs::RightLookupOperand } ) == ( { JoltR1CSInputs::RightInstructionInput } )
+    ),
+    // if Assert {
+    //     assert!(LookupOutput == 1)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::Assert) } }
+        => ( { JoltR1CSInputs::LookupOutput } ) == ( { 1i128 } )
+    ),
+    // if Rd != 0 && WriteLookupOutputToRD {
+    //     assert!(RdWriteValue == LookupOutput)
+    // }
+    r1cs_prod!(
+        ({ JoltR1CSInputs::Rd })
+            * ({ JoltR1CSInputs::OpFlags(CircuitFlags::WriteLookupOutputToRD) })
+            == ({ JoltR1CSInputs::WriteLookupOutputToRD })
+    ),
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::WriteLookupOutputToRD } }
+        => ( { JoltR1CSInputs::RdWriteValue } ) == ( { JoltR1CSInputs::LookupOutput } )
+    ),
+    // if Rd != 0 && Jump {
+    //     if !isCompressed {
+    //          assert!(RdWriteValue == UnexpandedPC + 4)
+    //     } else {
+    //          assert!(RdWriteValue == UnexpandedPC + 2)
+    //     }
+    // }
+    r1cs_prod!(
+        ({ JoltR1CSInputs::Rd }) * ({ JoltR1CSInputs::OpFlags(CircuitFlags::Jump) })
+            == ({ JoltR1CSInputs::WritePCtoRD })
+    ),
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::WritePCtoRD } }
+        => ( { JoltR1CSInputs::RdWriteValue } ) == ( { JoltR1CSInputs::UnexpandedPC } + { 4i128 } - { 2 * JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed) } )
+    ),
+    // if Jump && !NextIsNoop {
+    //     assert!(NextUnexpandedPC == LookupOutput)
+    // }
+    r1cs_prod!(
+        ({ JoltR1CSInputs::OpFlags(CircuitFlags::Jump) })
+            * ({ 1i128 } - { JoltR1CSInputs::NextIsNoop })
+            == ({ JoltR1CSInputs::ShouldJump })
+    ),
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::ShouldJump } }
+        => ( { JoltR1CSInputs::NextUnexpandedPC } ) == ( { JoltR1CSInputs::LookupOutput } )
+    ),
+    // if Branch && LookupOutput {
+    //     assert!(NextUnexpandedPC == UnexpandedPC + Imm)
+    // }
+    r1cs_prod!(
+        ({ JoltR1CSInputs::OpFlags(CircuitFlags::Branch) }) * ({ JoltR1CSInputs::LookupOutput })
+            == ({ JoltR1CSInputs::ShouldBranch })
+    ),
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::ShouldBranch } }
+        => ( { JoltR1CSInputs::NextUnexpandedPC } ) == ( { JoltR1CSInputs::UnexpandedPC } + { JoltR1CSInputs::Imm } )
+    ),
+    // if !(ShouldBranch || Jump) {
+    //     if DoNotUpdatePC {
+    //         assert!(NextUnexpandedPC == UnexpandedPC)
+    //     } else if isCompressed {
+    //         assert!(NextUnexpandedPC == UnexpandedPC + 2)
+    //     } else {
+    //         assert!(NextUnexpandedPC == UnexpandedPC + 4)
+    //     }
+    // }
+    // Note that ShouldBranch and Jump instructions are mutually exclusive
+    r1cs_prod!(
+        ({ JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed) })
+            * ({ JoltR1CSInputs::OpFlags(CircuitFlags::DoNotUpdateUnexpandedPC) })
+            == ({ JoltR1CSInputs::CompressedDoNotUpdateUnexpPC })
+    ),
+    r1cs_eq_conditional!(
+        if { { 1i128 } - { JoltR1CSInputs::ShouldBranch } - { JoltR1CSInputs::OpFlags(CircuitFlags::Jump) } }
+        => ( { JoltR1CSInputs::NextUnexpandedPC } )
+           == ( { JoltR1CSInputs::UnexpandedPC } + { 4i128 }
+                - { 4 * JoltR1CSInputs::OpFlags(CircuitFlags::DoNotUpdateUnexpandedPC) }
+                - { 2 * JoltR1CSInputs::OpFlags(CircuitFlags::IsCompressed) }
+                + { 2 * JoltR1CSInputs::CompressedDoNotUpdateUnexpPC } )
+    ),
+    // if Inline {
+    //     assert!(NextPC == PC + 1)
+    // }
+    r1cs_eq_conditional!(
+        if { { JoltR1CSInputs::OpFlags(CircuitFlags::InlineSequenceInstruction) } }
+        => ( { JoltR1CSInputs::NextPC } ) == ( { JoltR1CSInputs::PC } + { 1i128 } )
+    ),
+];
