+++ jolt-core/src/zkvm/r1cs/mod.rs
@@ -1,8 +1,12 @@
-pub mod inputs;
-
-pub mod builder;
-pub mod const_constraints;
 pub mod constraints;
+pub mod inputs;
 pub mod key;
 pub mod ops;
 pub mod spartan;
+
+#[cfg(test)]
+pub mod old_builder;
+#[cfg(test)]
+pub mod old_constraints;
+#[cfg(test)]
+pub mod old_ops;
