+++ jolt-core/src/zkvm/r1cs/spartan.rs
@@ -16,13 +16,10 @@ use crate::utils::math::Math;
 use crate::zkvm::dag::stage::SumcheckStages;
 use crate::zkvm::dag::state_manager::{ProofData, ProofKeys, StateManager};
 use crate::zkvm::instruction::CircuitFlags;
-use crate::zkvm::r1cs::builder::Constraint;
-use crate::zkvm::r1cs::constraints::{JoltRV32IMConstraints, R1CSConstraints};
-use crate::zkvm::r1cs::inputs::JoltR1CSInputs;
-use crate::zkvm::r1cs::inputs::ALL_R1CS_INPUTS;
-use crate::zkvm::r1cs::inputs::COMMITTED_R1CS_INPUTS;
+use crate::zkvm::r1cs::constraints::UNIFORM_R1CS;
+use crate::zkvm::r1cs::inputs::{JoltR1CSInputs, ALL_R1CS_INPUTS, COMMITTED_R1CS_INPUTS};
 use crate::zkvm::r1cs::key::UniformSpartanKey;
-use crate::zkvm::witness::{CommittedPolynomial, VirtualPolynomial};
+use crate::zkvm::witness::VirtualPolynomial;
 
 use crate::transcripts::Transcript;
 use crate::utils::small_value::NUM_SVO_ROUNDS;
@@ -35,9 +32,6 @@ use crate::{
     poly::{dense_mlpoly::DensePolynomial, eq_poly::EqPlusOnePolynomial},
     subprotocols::sumcheck::{SumcheckInstance, SumcheckInstanceProof},
 };
-
-use super::builder::CombinedUniformBuilder;
-
 use rayon::prelude::*;
 
 #[derive(Clone, Debug)]
@@ -91,22 +85,15 @@ where
     ProofTranscript: Transcript,
 {
     #[tracing::instrument(skip_all, name = "Spartan::setup")]
-    pub fn setup(
-        constraint_builder: &CombinedUniformBuilder<F>,
-        padded_num_steps: usize,
-    ) -> UniformSpartanKey<F> {
-        assert_eq!(
-            padded_num_steps,
-            constraint_builder.uniform_repeat().next_power_of_two()
-        );
-        UniformSpartanKey::from_builder(constraint_builder)
+    #[inline]
+    pub fn setup(padded_num_steps: usize) -> UniformSpartanKey<F> {
+        UniformSpartanKey::new(padded_num_steps)
     }
 
     #[tracing::instrument(skip_all)]
     fn prove_outer_sumcheck(
         num_rounds_x: usize,
         uniform_constraints_only_padded: usize,
-        uniform_constraints: &[Constraint],
         input_polys: &[MultilinearPolynomial<F>],
         tau: &[F],
         transcript: &mut ProofTranscript,
@@ -114,7 +101,6 @@ where
         SumcheckInstanceProof::prove_spartan_small_value::<NUM_SVO_ROUNDS>(
             num_rounds_x,
             uniform_constraints_only_padded,
-            uniform_constraints,
             input_polys,
             tau,
             transcript,
@@ -168,16 +154,20 @@ impl<F: JoltField> InnerSumcheck<F> {
             .for_each(|(r1cs_input, dest)| {
                 let accumulator = state_manager.get_prover_accumulator();
                 let accumulator = accumulator.borrow();
+                let key = r1cs_input.to_opening_id().expect(
+                    "Failed to map R1CS input to OpeningId (neither virtual nor committed)",
+                );
                 let (_, claim) = accumulator
                     .openings
-                    .get(&r1cs_input.try_into().ok().unwrap())
-                    .unwrap();
+                    .get(&key)
+                    .expect("Missing opening claim for expected OpeningId in bind_z");
                 *dest = *claim;
             });
 
         // Set the constant value at the appropriate position
-        if key.uniform_r1cs.num_vars < num_vars_uniform {
-            bind_z[key.uniform_r1cs.num_vars] = F::one();
+        let const_col = JoltR1CSInputs::num_inputs();
+        if const_col < num_vars_uniform {
+            bind_z[const_col] = F::one();
         }
 
         drop(_guard);
@@ -615,9 +605,7 @@ pub struct SpartanDag<F: JoltField> {
 
 impl<F: JoltField> SpartanDag<F> {
     pub fn new<ProofTranscript: Transcript>(padded_trace_length: usize) -> Self {
-        let constraint_builder = JoltRV32IMConstraints::construct_constraints(padded_trace_length);
         let key = Arc::new(UniformSpartanProof::<F, ProofTranscript>::setup(
-            &constraint_builder,
             padded_trace_length,
         ));
         Self { key }
@@ -640,7 +628,6 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
         let (preprocessing, trace, _program_io, _final_memory_state) =
             state_manager.get_prover_data();
 
-        let padded_trace_length = trace.len().next_power_of_two();
         let key = self.key.clone();
 
         // Create input polynomials from trace
@@ -656,22 +643,14 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
             .borrow_mut()
             .challenge_vector(num_rounds_x);
 
-        // Recreate constraint_builder from padded_trace_length
-        let constraint_builder: CombinedUniformBuilder<F> =
-            JoltRV32IMConstraints::construct_constraints(padded_trace_length);
-
-        let uniform_constraints_only_padded = constraint_builder
-            .uniform_builder
-            .constraints
-            .len()
-            .next_power_of_two();
+        // Uniform constraints per-step (padded to power-of-two) from constants
+        let uniform_constraints_only_padded = UNIFORM_R1CS.len().next_power_of_two();
 
         let (outer_sumcheck_proof, outer_sumcheck_r, outer_sumcheck_claims) = {
             let mut transcript = state_manager.transcript.borrow_mut();
             UniformSpartanProof::<F, ProofTranscript>::prove_outer_sumcheck(
                 num_rounds_x,
                 uniform_constraints_only_padded,
-                &constraint_builder.uniform_builder.constraints,
                 &input_polys,
                 &tau,
                 &mut transcript,
@@ -727,7 +706,7 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
         // proven using the PCS opening proof, which we add for future opening proof here
         let committed_polys: Vec<_> = COMMITTED_R1CS_INPUTS
             .iter()
-            .map(|input| CommittedPolynomial::try_from(*input).ok().unwrap())
+            .map(|input| input.to_committed_polynomial().ok().unwrap())
             .collect();
         let committed_poly_claims: Vec<_> = COMMITTED_R1CS_INPUTS
             .iter()
@@ -748,7 +727,7 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
             // Skip if it's a committed input (already added above)
             if !COMMITTED_R1CS_INPUTS.contains(input) {
                 accumulator.borrow_mut().append_virtual(
-                    VirtualPolynomial::try_from(*input).ok().unwrap(),
+                    input.to_virtual_polynomial().ok().unwrap(),
                     SumcheckId::SpartanOuter,
                     OpeningPoint::new(r_cycle.to_vec()),
                     *eval,
@@ -852,7 +831,7 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
         // proven using the PCS opening proof, which we add for future opening proof here
         let committed_polys: Vec<_> = COMMITTED_R1CS_INPUTS
             .iter()
-            .map(|input| CommittedPolynomial::try_from(*input).ok().unwrap())
+            .map(|input| input.to_committed_polynomial().ok().unwrap())
             .collect();
         accumulator.borrow_mut().append_dense(
             committed_polys,
@@ -864,7 +843,7 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
             // Skip if it's a committed input (already added above)
             if !COMMITTED_R1CS_INPUTS.contains(input) {
                 accumulator.borrow_mut().append_virtual(
-                    VirtualPolynomial::try_from(*input).ok().unwrap(),
+                    input.to_virtual_polynomial().ok().unwrap(),
                     SumcheckId::SpartanOuter,
                     OpeningPoint::new(r_cycle.to_vec()),
                 );
@@ -953,7 +932,7 @@ impl<F: JoltField, ProofTranscript: Transcript, PCS: CommitmentScheme<Field = F>
                 let accumulator = accumulator.borrow();
                 let (_, claim) = accumulator
                     .openings
-                    .get(&r1cs_input.try_into().ok().unwrap())
+                    .get(&r1cs_input.to_opening_id().ok().unwrap())
                     .unwrap();
                 *claim
             })
